refactor: consolidate ginkgo wrappers into one file
