--- test/e2e/framework/ginkgowrapper/BUILD
@@ -1,26 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-)
-
-go_library(
-    name = "go_default_library",
-    srcs = ["wrapper.go"],
-    importpath = "k8s.io/kubernetes/test/e2e/framework/ginkgowrapper",
-    deps = ["//vendor/github.com/onsi/ginkgo:go_default_library"],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-)
